Scrolling the div contents with Left and Right Scroll Buttons
If we have a requirement to scroll a div container horizontally without
using the scrollbar, then this is the apt one. The script is going to
detect the div content overflow and will automatically add the left
scroll and right scroll buttons on either side of the main div
container. User can click on these buttons to scroll right or left.